requirements.txt
Django>=5,<6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
whitenoise
dj-database-url
python-dotenv
psycopg2-binary  
gunicorn
openai
